Fix action list overflowing on Firefox
This may not be desirable, since it hardcodes the 84px gap needed to ensure that the top and bottom toolbars fit (30 px height + 5 px padding on each side, and 1px border on each side, times 2 toolbars). But it does fix the inspector pane taking up the full height on Firefox. I don't know why it works fine in other browsers without this though...